change the `optype.infer`  module into a package
